requirements.txt
streamlit==1.35.0
torch>=2.0.0
torchvision
Pillow
numpy
faiss-cpu
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
